Start laying out the Components class
